remove inccompilation = null; add default support to non-inccomp; native lib wouldn't be compiled twice
