+++ sandbox/src/main/scala/brainfuck/BrainfuckBenchmark.scala
@@ -2,7 +2,8 @@
 // Copyright (c) 2014 'Konstantin Makarchev'
 //
 //    MIT License
-//
+
+
 //    Permission is hereby granted, free of charge, to any person obtaining
 //    a copy of this software and associated documentation files (the
 //    "Software"), to deal in the Software without restriction, including
