+++ sandbox/src/main/scala/mylib.scala
@@ -1,0 +1,5 @@
+class mylib {
+  def print(): Unit = {
+    println("hellow11")
+  }
+}
